fix(examples/chat): stage recording validator types its kinds and checks checkpointIndex and histories; fixture moves out of the spec
