+++ examples/chat/angular/src/app/stage/stage-recording.types.spec.ts
@@ -1,24 +1,7 @@
 // examples/chat/angular/src/app/stage/stage-recording.types.spec.ts
 import { describe, expect, it } from 'vitest';
-import { validateStageRecording, type StageRecording } from './stage-recording.types';
-
-const ev = (tMs: number) => ({ tMs, event: { type: 'values', messages: [] } as never });
-
-export const MINIMAL: StageRecording = {
-  version: 2,
-  recordedAt: '2026-09-06T00:00:00.000Z',
-  threadId: 'thread-1',
-  runs: [
-    { beat: 'stream', action: { kind: 'submit', message: 'Tell me about signals' }, events: [ev(0), ev(50)] },
-    { beat: 'persist', action: { kind: 'reload' }, events: [] },
-    { beat: 'persist', action: { kind: 'submit', message: 'Shorter, please.' }, events: [ev(0)] },
-    { beat: 'persist', action: { kind: 'submit', message: 'As a haiku.', checkpointIndex: 1 }, events: [ev(0)] },
-    { beat: 'approve', action: { kind: 'submit', message: 'Clean up backups.' }, events: [ev(0), ev(900)] },
-    { beat: 'approve', action: { kind: 'resume', value: 'approved' }, events: [ev(0)] },
-    { beat: 'render', action: { kind: 'submit', message: 'Show me a form.' }, events: [ev(0), ev(2000)] },
-  ],
-  histories: [{ afterRun: 1, states: [] }],
-};
+import { validateStageRecording } from './stage-recording.types';
+import { MINIMAL } from './stage-recording.fixtures';
 
 describe('validateStageRecording', () => {
   it('accepts a well-formed recording', () => {
@@ -39,5 +22,16 @@ describe('validateStageRecording', () => {
   it('requires a resume run to follow the approve submit', () => {
     const bad = { ...MINIMAL, runs: MINIMAL.runs.filter((r) => r.action.kind !== 'resume') };
     expect(() => validateStageRecording(bad)).toThrow(/resume/);
+  });
+  it('rejects a negative checkpointIndex', () => {
+    const bad = {
+      ...MINIMAL,
+      runs: MINIMAL.runs.map((r, i) => (i === 3 ? { ...r, action: { ...r.action, checkpointIndex: -1 } } : r)),
+    };
+    expect(() => validateStageRecording(bad)).toThrow(/checkpointIndex/);
+  });
+  it('rejects a history with a non-numeric afterRun', () => {
+    const bad = { ...MINIMAL, histories: [{ afterRun: 'x', states: [] }] };
+    expect(() => validateStageRecording(bad)).toThrow(/afterRun/);
   });
 });
